Updated vehicle status enum. Added status check for delete vehicle API.

+++ src/main/java/com/fa20se28/vma/component/impl/VehicleComponentImpl.java
@@ -2,6 +2,7 @@ package com.fa20se28.vma.component.impl;
 
 import com.fa20se28.vma.component.VehicleComponent;
 import com.fa20se28.vma.configuration.exception.DataException;
+import com.fa20se28.vma.enums.VehicleStatus;
 import com.fa20se28.vma.mapper.*;
 import com.fa20se28.vma.model.Vehicle;
 import com.fa20se28.vma.model.VehicleDetail;
@@ -48,7 +49,7 @@ public class VehicleComponentImpl implements VehicleComponent {
 
     @Override
     public List<Vehicle> getVehicles(VehiclePageReq request, int viewOption, int pageNum, String ownerId) {
-        return vehicleMapper.getVehicles(request, viewOption, pageNum*15, ownerId);
+        return vehicleMapper.getVehicles(request, viewOption, pageNum * 15, ownerId);
     }
 
     @Override
@@ -61,8 +62,7 @@ public class VehicleComponentImpl implements VehicleComponent {
     public void assignVehicle(String vehicleId, String driverId) {
         int row = issuedVehicleMapper.assignVehicle(vehicleId, driverId);
 
-        if (row == 0)
-        {
+        if (row == 0) {
             throw new DataException("Unknown error occurred. Data not added!");
         }
     }
@@ -72,8 +72,7 @@ public class VehicleComponentImpl implements VehicleComponent {
     public void withdrawVehicle(String vehicleId) {
         int row = issuedVehicleMapper.withdrawVehicle(vehicleId);
 
-        if (row == 0)
-        {
+        if (row == 0) {
             throw new DataException("Unknown error occurred. Data not modified!");
         }
     }
@@ -81,64 +80,52 @@ public class VehicleComponentImpl implements VehicleComponent {
     @Override
     @Transactional
     public void createVehicle(VehicleReq vehicle) {
-        if (!vehicleMapper.isVehicleExist(vehicle.getVehicleId()))
-        {
+        if (!vehicleMapper.isVehicleExist(vehicle.getVehicleId())) {
             int vehicleRow;
-            if (vehicle.getRoleId() == 2)
-            {
-                vehicleRow = vehicleMapper.createVehicle(vehicle, "PENDING_APPROVAL");
-            }
-            else
-            {
-                vehicleRow = vehicleMapper.createVehicle(vehicle, "NO_DRIVER");
+            if (vehicle.getRoleId() == 2) {
+                vehicleRow = vehicleMapper.createVehicle(vehicle, VehicleStatus.PENDING_APPROVAL);
+            } else {
+                vehicleRow = vehicleMapper.createVehicle(vehicle, VehicleStatus.AVAILABLE_NO_DRIVER);
             }
             int documentRowCount = 0;
             int documentImageRowCount = 0;
-            if (vehicleRow != 0)
-            {
-                for (VehicleDocumentReq doc: vehicle.getVehicleDocuments())
-                {
-                    if (!vehicleDocumentMapper.isDocumentExist(doc.getVehicleDocumentId()))
-                    {
+            if (vehicleRow != 0) {
+                for (VehicleDocumentReq doc : vehicle.getVehicleDocuments()) {
+                    if (!vehicleDocumentMapper.isDocumentExist(doc.getVehicleDocumentId())) {
                         int vehicleDoc = vehicleDocumentMapper.createVehicleDocument(doc, vehicle.getVehicleId());
-                        if (vehicleDoc != 0)
-                        {
-                            for (String image: doc.getImageLink())
-                            {
+                        if (vehicleDoc != 0) {
+                            for (String image : doc.getImageLink()) {
                                 int vehicleDocImage = vehicleDocumentImageMapper.createVehicleDocumentImage(doc.getVehicleDocumentId(), image);
-                                if (vehicleDocImage != 0)
-                                {
+                                if (vehicleDocImage != 0) {
                                     documentImageRowCount++;
                                 }
                             }
                             documentRowCount++;
                         }
-                    }
-                    else
-                    {
-                        throw new DataException("Document with ID "+ doc.getVehicleDocumentId() + " already exist!");
+                    } else {
+                        throw new DataException("Document with ID " + doc.getVehicleDocumentId() + " already exist!");
                     }
                 }
             }
 
-            if (vehicleRow == 0 && documentRowCount == 0 && documentImageRowCount == 0)
-            {
+            if (vehicleRow == 0 && documentRowCount == 0 && documentImageRowCount == 0) {
                 throw new DataException("Unknown error occurred. Data not added!");
             }
-        }
-        else
-        {
+        } else {
             throw new DataException("Vehicle with ID " + vehicle.getVehicleId() + " already exist!");
         }
     }
 
     @Override
     public void deleteVehicle(String vehicleId) {
-        int row = vehicleMapper.deleteVehicle(vehicleId);
+        if (vehicleMapper.getVehicleStatus(vehicleId) != VehicleStatus.DELETED) {
+            int row = vehicleMapper.deleteVehicle(vehicleId);
 
-        if (row == 0)
-        {
-            throw new DataException("Unknown error occurred. Data not modified!");
+            if (row == 0) {
+                throw new DataException("Unknown error occurred. Data not modified!");
+            }
+        } else {
+            throw new DataException("Vehicle is already deleted!");
         }
     }
